chuyển qua dùng dạng tool cho nó dễ xử lý
